requirements.txt
Django==5.0
djangorestframework==3.14.0
Pillow==10.1.0
drf-yasg==1.21.7
requests==2.31.0
google-auth==2.25.2
PyJWT==2.8.0
